requirements.txt
flask
flask-cors
pypdf
pikepdf
pymupdf
python-pptx
pdf2docx
img2pdf
pillow
pdfminer.six
pdfplumber
gunicorn
celery
redis
pytesseract
pandas
openpyxl
reportlab